Database and JTable connection and other things
-Linked together the database and the Table so now songs added will be
seen(Need to restart the application for it to take affect, this will
need fixing in the future it isn't very important right now)
-Added functionality to add a song to the database through the "Add a
song" menu button. Pressing "Add a song" will open up a file explorer
and you can add songs to the database.
-The JTable is now responsive! Selecting a different item will set the
new file path for that song so now you can press play and it will play
the correct song. Play button will need to be pressed twice for some
reason - needs future fix but at least it plays the song.
-Switched to PreparedStatements for the insertion of the file paths for
the MP3 files