Delete Menu Button, small bug and visual fixes
Added functionality to the delete song standard menu button.

Changed openFileExplorer from a void to a boolean and changed its name
to openFileExplorerAndAddedSong so that when a user opens up the file
explorer but closes it before they pick a song an empty object won't get
added to the database.

Also changed it in the function addsongToTable so that when adding an
item to the table it will go through dispNull to make it look better.