Right click add songs to playlist + other stuff
Users can now right click on a song in the main song table view and add
it to a playlist. This adds the song to the database and if the user
clicks on the playlist the added songs will appear.

-Made it so the use can just press enter after entering the name of
their playlist

-Fixed the error that when playlist are created the genre field was
still an integer. This was fixed.

-Made it so that the song library/playlist and the hierarchy tree can be
resizable with the introduction of JSplitPane.

-Implemented the popup menu for the hierarchy panel with "Delete
Playlist" and "Open in a new Window" buttons. Buttons need
implementation.